python-system/requirements.txt
opencv-python
ultralytics
pyttsx3
customtkinter
pillow
